// src/types/api/reservation-history.ts
export type ReservationHistoryItem = {
  HistoryID?: string
  HistoryDt?: string | null
  HistoryAction?: string | null
  ReservationID?: string | null
  CustomerID?: string | null
  BusinessID?: string | null
  ShowID?: string | null
  ShowDetID?: string | null
  Dinner?: string | null
  OrigPartyNo?: number | null
  PartyNo?: number | null
  ResSec?: string | null
  TableNums?: string | null
  TixPaid?: number | null
  TixComp?: number | null
  TixDisc?: number | null
  CheckedIn?: number | null
  CheckInPaid?: number | null
  CheckInComp?: number | null
  CheckInDisc?: number | null
  Printed?: string | null
  VIP?: string | null
  Price?: number | null
  DayOfShowCharge?: number | null
  PhoneCharge?: number | null
  WalkupCharge?: number | null
  WebCharge?: number | null
  Promo?: string | null
  NumPasses?: number | null
  PendingStatus?: string | null
  SubTot?: number | null
  SVC?: number | null
  Discount?: number | null
  SalesTax?: number | null
  Total?: number | null
  ResSource?: string | null
  ResStatus?: string | null
  ResTemp?: string | null
  ActionForm?: string | null
  Action?: string | null
  CreatedBy?: string | null
  CreateDt?: string | null
  CancelBy?: string | null
  CancelDt?: string | null
  UpdateCount?: number | null
  LastUpdateID?: string | null
  LastUpdateDt?: string | null
}