// src/types/reservation-history.ts
export type ReservationHistoryRow = {
  historyId: string
  historyDate: string
  historyAction: string
  reservationId: string
  customerId: string
  showId: string
  showDetId: string
  dinner: string
  origPartyNo: string
  partyNo: string
  resSec: string
  tableNums: string
  tixPaid: string
  tixComp: string
  tixDisc: string
  checkedIn: string
  checkInPaid: string
  checkInComp: string
  checkInDisc: string
  printed: string
  vip: string
  price: string
  dayOfShowCharge: string
  phoneCharge: string
  walkupCharge: string
  webCharge: string
  promo: string
  numPasses: string
  pendingStatus: string
  subTot: string
  svc: string
  discount: string
  salesTax: string
  total: string
  resSource: string
  resStatus: string
  resTemp: string
  actionForm: string
  action: string
  createdBy: string
  createDt: string
  cancelBy: string
  cancelDt: string
  updateCount: string
  lastUpdateId: string
  lastUpdateDt: string
}